Move RenderTask and related struct to their own file
This is another commit which addresses #587.
