+++ webrender/src/frame_builder.rs
@@ -9,31 +9,29 @@ use gpu_store::GpuStoreAddress;
 use internal_types::SourceTexture;
 use mask_cache::{ClipSource, MaskCacheInfo};
 use prim_store::{BorderPrimitiveCpu, BorderPrimitiveGpu, BoxShadowPrimitiveGpu};
-use prim_store::{GradientPrimitiveCpu, GradientPrimitiveGpu, ImagePrimitiveCpu};
-use prim_store::{ImagePrimitiveGpu, ImagePrimitiveKind, PrimitiveContainer, PrimitiveGeometry};
-use prim_store::{PrimitiveIndex, PrimitiveStore, RadialGradientPrimitiveCpu};
-use prim_store::{RadialGradientPrimitiveGpu, RectanglePrimitive, TextRunPrimitiveCpu};
-use prim_store::{TextRunPrimitiveGpu, YuvImagePrimitiveCpu, YuvImagePrimitiveGpu};
+use prim_store::{GradientPrimitiveCpu, GradientPrimitiveGpu, ImagePrimitiveCpu, ImagePrimitiveGpu};
+use prim_store::{ImagePrimitiveKind, PrimitiveContainer, PrimitiveGeometry, PrimitiveIndex};
+use prim_store::{PrimitiveStore, RadialGradientPrimitiveCpu, RadialGradientPrimitiveGpu};
+use prim_store::{RectanglePrimitive, TextRunPrimitiveCpu, TextRunPrimitiveGpu};
+use prim_store::{YuvImagePrimitiveCpu, YuvImagePrimitiveGpu};
 use profiler::FrameProfileCounters;
+use render_task::{AlphaRenderItem, MaskCacheKey, MaskResult, RenderTask, RenderTaskIndex};
+use render_task::RenderTaskLocation;
 use resource_cache::ResourceCache;
 use scroll_tree::ScrollTree;
-use std::cmp;
-use std::mem;
-use std::usize;
-use std::{i32, f32};
-use tiling::{AlphaRenderItem, AuxiliaryListsMap, CompositeOps, Frame, MaskCacheKey, MaskResult};
-use tiling::{PackedLayer, PackedLayerIndex, PrimitiveFlags, PrimitiveRunCmd, RenderPass};
-use tiling::{RenderTargetContext, RenderTask, RenderTaskCollection, RenderTaskIndex};
-use tiling::{RenderTaskLocation, ScrollLayer, ScrollLayerIndex, ScrollbarPrimitive};
+use std::{cmp, f32, i32, mem, usize};
+use tiling::{AuxiliaryListsMap, CompositeOps, Frame, PackedLayer, PackedLayerIndex};
+use tiling::{PrimitiveFlags, PrimitiveRunCmd, RenderPass, RenderTargetContext};
+use tiling::{RenderTaskCollection, ScrollbarPrimitive, ScrollLayer, ScrollLayerIndex};
 use tiling::{StackingContext, StackingContextIndex};
-use util::{TransformedRect, TransformedRectKind};
-use util::{self, pack_as_float, rect_from_points_f, subtract_rect};
-use webrender_traits::{BorderDisplayItem, BorderSide, BorderStyle, BoxShadowClipMode, ClipRegion};
-use webrender_traits::{ColorF, DeviceIntPoint, DeviceIntRect, DeviceIntSize, DeviceUintSize};
-use webrender_traits::{ExtendMode, FontKey, FontRenderMode, GlyphOptions, ImageKey, ImageRendering};
-use webrender_traits::{ItemRange, LayerPoint, LayerRect, LayerSize, LayerToScrollTransform};
-use webrender_traits::{PipelineId, ScrollLayerId, ScrollLayerPixel, WebGLContextId, WorldRect};
-use webrender_traits::{YuvColorSpace, as_scroll_parent_rect, device_length};
+use util::{self, pack_as_float, rect_from_points_f, subtract_rect, TransformedRect};
+use util::TransformedRectKind;
+use webrender_traits::{as_scroll_parent_rect, BorderDisplayItem, BorderSide, BorderStyle};
+use webrender_traits::{BoxShadowClipMode, ClipRegion, ColorF, device_length, DeviceIntPoint};
+use webrender_traits::{DeviceIntRect, DeviceIntSize, DeviceUintSize, ExtendMode, FontKey};
+use webrender_traits::{FontRenderMode, GlyphOptions, ImageKey, ImageRendering, ItemRange};
+use webrender_traits::{LayerPoint, LayerRect, LayerSize, LayerToScrollTransform, PipelineId};
+use webrender_traits::{ScrollLayerId, ScrollLayerPixel, WebGLContextId, WorldRect, YuvColorSpace};
 
 #[derive(Clone, Copy)]
 pub struct FrameBuilderConfig {
